requirements.txt
amqp==5.0.6
aniso8601==7.0.0
asgiref==3.3.1
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
distlib==0.3.3
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-graphql-playground==1.0
django-redis==5.0.0
django-seed==0.2.2
django-sendgrid-v5==0.9.1
djangorestframework==3.12.2
environ==1.0
Faker==8.0.0
filelock==3.3.0
future==0.18.2
gitdb==4.0.7
gitflow==0.5.1
GitPython==3.1.14
graphene==2.1.8
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
kombu==5.1.0
platformdirs==2.4.0
promise==2.3
prompt-toolkit==3.0.20
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-http-client==3.3.2
pytz==2021.1
redis==3.5.3
Rx==1.6.1
sendgrid==6.7.0
singledispatch==3.6.1
six==1.15.0
smmap==4.0.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
text-unidecode==1.3
vine==5.0.0
virtualenv==20.8.1
wcwidth==0.2.5
whitenoise==5.2.0
